requirements.txt
asgiref==3.4.1
astroid==2.9.0
backports.zoneinfo==0.2.1
Django==4.0
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.10.0
isort==5.10.1
lazy-object-proxy==1.6.0
Markdown==3.3.6
mccabe==0.6.1
platformdirs==2.4.0
pylint==2.12.2
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.0.1
wrapt==1.13.3
zipp==3.6.0
